Get char[] data directly from ContentLine Use String to drawText instead of char[]

+++ editor/src/main/java/io/github/rosemoe/sora/graphics/FontCache.java
@@ -57,7 +57,7 @@ public class FontCache {
      * Measure a single character
      */
     public float measureChar(char ch, Paint p) {
-        float width = cache[(int) ch];
+        float width = cache[ch];
         if (width == 0) {
             buffer[0] = ch;
             width = p.measureText(buffer, 0, 1);
